added error handler added error message view
todo:
google sign-in
