+++ app/src/main/java/com/koray/nrcnewsapp/core/network/viewmodel/LiveCategoriesModel.kt
@@ -6,10 +6,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.koray.nrcnewsapp.core.domain.CategoryItemModel
 import com.koray.nrcnewsapp.core.network.dto.CategoryDto
+import com.koray.nrcnewsapp.core.network.helper.ErrorHandler
 import com.koray.nrcnewsapp.core.network.repository.CategoryRepository
 import com.koray.nrcnewsapp.core.network.transformer.CategoryTransformer
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.schedulers.Schedulers
+import java.lang.StringBuilder
 import javax.inject.Inject
 
 class LiveCategoriesModel @Inject constructor(
@@ -29,7 +31,9 @@ class LiveCategoriesModel @Inject constructor(
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe(
                 { response -> this.categoriesLiveData.value = this.transform(response) },
-                { error -> println("Error: this is my error!, ${error.message}") }
+                { error ->
+                    ErrorHandler.showError(StringBuilder(error.message!!))
+                }
             )
     }
